requirements.txt
gradio==3.50.2
transformers==4.37.2
torch==2.2.0
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
python-dotenv==1.0.1
huggingface-hub==0.20.3
pydantic==2.5.3
tavily-python==0.3.1 